Add the `cut_body_after` configuration option
This configuration option must be set to a string.
When this string is present, bors-ng will search the bodies of all PRs,
before generating the body for its merge commit.
If it is found, bors will not include the string or anything after it.

For example, if a pull request has a body like this:

> ### Synchronize background and foreground processing
>
> Fixes that annoying bug.
>
> <!-- boilerplate follows -->
>
> Thank you for contributing to my awesome OSS project!
> To make sure your PR is accepted ASAP, make sure all of this
> stuff is done:
>
> - [ ] Run the linter
> - [ ] Run any new or changed tests
> - [ ] This PR fixes #___ (fill in if it exists)
> - [ ] Make sure your commit messages make sense

And if the bors.toml file has this:

    cut_body_after = "\n<!-- boilerplate follows -->"

Then the resulting commit message will look like this:

> Merge #NN #MM r=bill
>
> NN: Synchronize background and foreground processing
>
> Fixes that annoying bug.

---

Fixes #200