WIP: close PRs after merge
Currently GitHub automatically closes the PR, once it detects that
master has received the merge commit. So this operation should be a
noop and is mainly there to ensure the right behaviour from GitHub.

In the future this will be useful to guarantee PRs are getting closed
with different merging strategies that are not detected by GitHub.