Add production Dockerfile using compiled releases
The previously existing Dockerfile in the scripts/ directory works for
development, but for production Bors-NG deployments it would be much
better to run from a compiled release, to reduce image size and startup
overhead.

Implemenet such an image using multi-stage Docker builds, as introduced
in Docker 17.05. The builder stage contains the Erlang, Elixir and
NodeJS runtimes, and is only used to generate a compiled release, which
is copied to a basic Debian Jessie image afterwards.