Prevent unwanted space after queue log PR links
Right now the spaces in between PR links in the log are also links, because there's extra whitespace in the `a` tag here. If we remove that extra whitespace in the `a` tag, then we should still have a space between links because of the indentation before the `a` tag giving us our needed whitespace, but it won't be inside the `a` tag so it won't be part of the link.