Don't simply ignore unset commit statuses
Fixes #1001.

Previously we treated unset statuses as :ok. This, I think, is
non-intuitive as it's entirely possible for CI to be slow to set even a
pending status. If someone `bors r+` before that's done, bors would go
ahead and create the merge commit, which is not the contract defined by
the `pr_status` config, which enumerates the list of commit statuses
that must pass on the PR commit when it is r+-ed.