Add custom commit titles
This commit adds the ability to customise commit titles using a new
config property: `commit_title`.

To maintain backwards compatibility, the commit_title defaults to
`Merge ${PR_REFS}`. The hardcoded token `${PR_REFS}` is replaced with
the available patchlinks' xrefs.

The commit_title can be configured in the Bors Toml.