Fix CODEOWNERS failure due to hardcoded "master"
Our repository does not use `master` as the mainline for bors. Bors would erroneously pass the codeowners check, and in the logs it would get a 404: 
```
[error] GET https://api.github.com/repositories/123456/contents/.github/CODEOWNERS -> 404 (127.281 ms)
```

*I have not tested this PR*. I do not know Elixir, but this seems to be a good starting point for a change. And improvement could be to fail if `use_codeowners` is true and the http request fails.

Hopefully this is a good jumping off point. My coworker knows Elixir and has contributed in the past, so maybe they can get this PR into a better state if it's not acceptable as-is :)