Load database repo dynamically
When creating an elixir release, the database adapter is hardcoded
into `sys.config`. With the hardcoded database, the docker image
is strictly set to Postgres.

This change allows to set the database url to a MySQL host by
environment variable, so the current docker image can connect to
MySQL and use the appropiate Ecto Adapter.

Co-authored-by: Björn Brauer <bjoern.brauer@new-work.se>
Co-authored-by: Theodor Fiedler <theodor.fiedler@new-work.se>
Co-authored-by: Philipp Hinrichsen <philipp.hinrichsen@new-work.se>
Co-authored-by: Markus Wolf <markus.wolf@new-work.se>