Fix `sync_patch`'s use of `draft` field
The map returned by `Jason.decode` uses strings for keys and not atoms:

```elixir
>>> {:ok, json} = Jason.decode(~s({ "draft": true }))
{:ok, %{"draft" => true}}
>>> json
%{"draft" => true}
```

… which meant that `pr[:draft]` would always return `nil` and so
`bors-ng` would never detect that a PR was in the draft state:

```elixir
>>> json[:draft]
nil
>>> json["draft"]
true
```

The fix is to check for `pr["draft"]`, which fixes the problem.

This wasn't caught by the tests because the tests create a map with
atom keys, which didn't trigger this problem.